fix(jsii-pacmak): underscore arguments conflict in Java
The `_` identifier is treated as a keyword since Java 9, and we used to
escape it to `__`.

The problem is that but it is somewhat likely that people will use `_,
__, ___` as multiple indifferent arguments, and now the identifiers will
conflict. We have to pick something else.

Ideally we would look at the surrounding argument names and pick
something guaranteed to be unique, but unfortunately the code isn't
quite structured that way so we'll pick something unlikely to collide
instead. We escape it to `_under_` instead.

Changing from `__` -> `_under_` would be a breaking change if applied to
public property names, but most likely this will be used for function
parameters (unfortunately the code has been structured in such a way
that property and parameter names are strongly tied together, in a way
that would take more time to unwind than I care to invest right now),
where it doesn't matter.