fix(jsii-pacmak): Maven resolve-ranges command is missing repo settings
In https://github.com/aws/jsii/pull/5006 we added a command to make
Maven resolve all the wide open version ranges to the latest version,
under the motivation that Maven doesn't have a distinction between a
dependency declaration file and a lockfile, and ranges don't have the
effect that we want (but DO contribute to long build times).

In that PR we forgot to pass the settings file that has repository
configuration information.

Add that.